requirements.txt
# Environment
python-dotenv==1.0.1
fastapi==0.115.8
uvicorn==0.34.0

# Libraries
langchain-core==0.3.37
langchain-openai==0.3.6
langgraph==0.2.74
